Move group functions to non-empty module
